8825_core/integrations/mcp/mcp_template/server.py
#!/usr/bin/env python3
"""
8825 v3.0 MCP Server Template
Base implementation for focus-specific MCP servers
"""

import os
import json
from pathlib import Path
from flask import Flask, request, jsonify
from flask_cors import CORS

# Load environment variables
from dotenv import load_dotenv
load_dotenv()

# Load MCP configuration
def load_config():
    """Load MCP configuration from config.json"""
    config_path = Path(__file__).parent / "config.json"
    
    if not config_path.exists():
        raise FileNotFoundError(f"Config file not found: {config_path}")
    
    with open(config_path, 'r') as f:
        config = json.load(f)
    
    # Expand environment variables in config
    config = expand_env_vars(config)
    
    return config

def expand_env_vars(obj):
    """Recursively expand environment variables in config"""
    if isinstance(obj, dict):
        return {k: expand_env_vars(v) for k, v in obj.items()}
    elif isinstance(obj, list):
        return [expand_env_vars(item) for item in obj]
    elif isinstance(obj, str) and obj.startswith('${') and obj.endswith('}'):
        var_name = obj[2:-1]
        return os.getenv(var_name, obj)
    return obj

# Initialize Flask app
app = Flask(__name__)
CORS(app)

# Load configuration
config = load_config()

# Health check endpoint
@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'mcp_id': config.get('mcp_id'),
        'name': config.get('name'),
        'version': '3.0.0'
    })

# Query knowledge endpoint
@app.route('/query', methods=['POST'])
def query_knowledge():
    """
    Query knowledge base
    
    Request body:
    {
        "query": "search term",
        "limit": 10,
        "filters": {}
    }
    """
    data = request.get_json()
    query = data.get('query', '')
    limit = data.get('limit', 10)
    filters = data.get('filters', {})
    
    # TODO: Implement actual search logic
    # This is a placeholder
    results = {
        'query': query,
        'results': [],
        'total': 0,
        'mcp_id': config.get('mcp_id')
    }
    
    return jsonify(results)

# List endpoints
@app.route('/endpoints', methods=['GET'])
def list_endpoints():
    """List available endpoints for this MCP"""
    return jsonify({
        'mcp_id': config.get('mcp_id'),
        'endpoints': config.get('endpoints', []),
        'access_control': config.get('access_control', {})
    })

# Get MCP info
@app.route('/info', methods=['GET'])
def get_info():
    """Get MCP information"""
    return jsonify({
        'mcp_id': config.get('mcp_id'),
        'name': config.get('name'),
        'port': config.get('port'),
        'data_sources': config.get('data_sources', []),
        'version': '3.0.0'
    })

# Main entry point
if __name__ == '__main__':
    port = config.get('port', 8825)
    debug = os.getenv('DEBUG', 'false').lower() == 'true'
    
    print(f"Starting {config.get('name')} on port {port}")
    print(f"MCP ID: {config.get('mcp_id')}")
    print(f"Endpoints: {', '.join(config.get('endpoints', []))}")
    
    app.run(
        host='0.0.0.0',
        port=port,
        debug=debug
    )
